modularize cloudsync
